+++ src/Phrase/match/01-matchAll.js
@@ -1,7 +1,7 @@
 const failFast = require('./02-failFast')
 const tryMatch = require('./03-tryMatch')
 const postProcess = require('./04-postProcess')
-const syntax = require('../../Doc/match/syntax')
+const syntax = require('../../Doc/syntax')
 
 /**  returns a simple array of arrays */
 const matchAll = function (p, regs, matchOne = false) {